+++ packages/core/src/interactivity/index.tsx
@@ -1,3 +1,5 @@
 export * from './DataComponent'
 export * from './SimpleDataComponent'
+export * from './TooltipDataComponent'
 export * from './types'
+export * from './utils'